refactor: move the CA provider check out of the closure
Eighth group of #669. CAProviderTest moves into modules/api/resources_ca.py.

Closure complexity 282 -> 249; resources.py 2295 -> 2023 lines. Since #669
started: 559 -> 249, and 4040 -> 2023 — half the file.

It gets a module of its own rather than joining the settings group it sits
beside in the closure: at ~270 lines it is substantial on its own, and probing
whether a CA endpoint answers is a different concern from managing
configuration.

test_ca_provider_wiring_consistency reads the provider literals out of the
source, so it is repointed at the new module — and confirmed still live by
breaking one of those literals and watching it fail. That check is the second
source-reading ratchet this decomposition has moved; a repointed ratchet that
silently reads nothing would pass forever, which is worse than deleting it.
