+++ tests/test_api_group_modules_are_standalone.py
@@ -18,8 +18,10 @@ from flask_restx import Api, Resource
 from modules.api.resource_context import ApiContext
 from modules.api.resources_backup import create_backup_resources
 from modules.api.resources_cache import create_cache_resources
+from modules.api.resources_ca import create_ca_resources
 from modules.api.resources_health import create_health_resources
 from modules.api.resources_inventory import create_inventory_resources
+from modules.api.resources_settings import create_settings_resources
 from modules.api.resources_storage import create_storage_resources
 
 pytestmark = [pytest.mark.unit]
@@ -278,3 +280,25 @@ def test_the_inventory_group_applies_scope_through_the_shared_helper(api):
         f'Anything else means the scope helper is not being applied — which '
         f'would hand a restricted caller the whole inventory.'
     )
+
+
+def test_the_ca_group_builds_from_a_minimal_context(api):
+    resources = create_ca_resources(api, {'ca_test_config_model': MagicMock()},
+                                    _minimal_context(managers={}))
+    assert set(resources) == {'CAProviderTest'}
+    assert issubclass(resources['CAProviderTest'], Resource)
+
+
+def test_the_settings_group_builds_from_a_minimal_context(api):
+    models = {
+        'settings_model': MagicMock(),
+        'dns_providers_model': MagicMock(),
+    }
+    ctx = _minimal_context(settings=MagicMock(), dns=MagicMock(),
+                           audit=MagicMock(), managers={})
+    resources = create_settings_resources(api, models, ctx)
+
+    assert set(resources) == {
+        'Settings', 'DNSProviders', 'DNSAccounts', 'DNSAccountDetail'}
+    for name, cls in resources.items():
+        assert issubclass(cls, Resource), f'{name} is not a Resource'
